+++ gemfiles/common.rb
@@ -8,3 +8,6 @@ gem "phenix", ">= 1.0.1"
 gem "pry-byebug", "~> 3.9"
 gem "rake", ">= 12.0.0"
 gem "standard"
+
+gem "mysql2"
+gem "trilogy", ">= 2.5.0"