+++ lib/active_record_host_pool.rb
@@ -4,18 +4,16 @@ require "active_record"
 require "active_record/base"
 require "active_record/connection_adapters/abstract_adapter"
 
-module ActiveRecordHostPool
-  class << self
-    attr_accessor :loaded_db_adapter
-  end
+begin
+  require "mysql2"
+rescue LoadError
+  :noop
 end
 
-if Gem.loaded_specs.include?("mysql2")
-  require "mysql2"
-  ActiveRecordHostPool.loaded_db_adapter = :mysql2
-elsif Gem.loaded_specs.include?("trilogy")
+begin
   require "trilogy"
-  ActiveRecordHostPool.loaded_db_adapter = :trilogy
+rescue LoadError
+  :noop
 end
 
 require "active_record_host_pool/clear_query_cache_patch"