+++ Changelog.md
@@ -6,6 +6,9 @@ and as of v1.0.0 this project adheres to [Semantic Versioning](https://semver.or
 
 ## [Unreleased]
 
+### Changed
+- ActiveRecordHostPool can now work with `mysql2` and `trilogy` if both gems are loaded.
+
 ### Added
 - Testing with Rails 8.0.